// src/modules/solicitacao/enums/tipo-condicao-regra.enum.ts
export enum TipoCondicaoRegra {
  TIPO_VEICULO = 'tipo-veiculo',
  TIPO_CORRIDA = 'tipo-corrida',
  DISTANCIA_MINIMA_KM = 'distancia-minima-km',
  DISTANCIA_MAXIMA_KM = 'distancia-maxima-km',
  HORA_INICIO = 'hora-inicio',
  HORA_FIM = 'hora-fim',
  DIA_SEMANA = 'dia-semana',
  QUANTIDADE_MINIMA_PARADAS = 'quantidade-minima-paradas',
}